chore(deps): bump pytest 9, pytest-asyncio 1.4, ruff 0.15 (#118)
Closes the python-tooling Dependabot group. pytest 8.3.3->9.1.0 + pytest-asyncio 0.24->1.4: migrate to the native session loop-scope (asyncio_default_test_loop_scope=session); the old 0.24 event_loop-override hack is gone. pytest>=9 lands the GHSA-6w46-j5rx-g56g fix, so pip-audit drops --ignore-vuln (TD-29 closed; verified locally: 'No known vulnerabilities found'). ruff 0.7.4->0.15.17: safe fixes (C420 dict.fromkeys, RUF023 sorted __slots__) + format; fixed RUF046 (redundant int(round())) and RUF043 (raw match= pattern). Ignored UP042 (str-Enum->StrEnum) and UP046 (PEP-695 generics) — behaviour-changing modernizations deferred to a dedicated PR.

+++ backend/app/api/webhooks.py
@@ -480,7 +480,7 @@ async def webhook_metrics(
     success_rate = (row.succeeded / terminal) if terminal else None
 
     def _ms(v: float | None) -> int | None:
-        return int(round(v)) if v is not None else None
+        return round(v) if v is not None else None
 
     return WebhookDeliveryStats(
         window_days=window_days,
